change package position, merge lazycolumn
