+++ app/src/main/java/com/codingbot/algorithm/data/model/sorting/ISortingAlgorithm.kt
@@ -1,6 +1,7 @@
-package com.codingbot.algorithm.data.model
+package com.codingbot.algorithm.data.model.sorting
 
 import com.codingbot.algorithm.data.SortingData
+import com.codingbot.algorithm.data.model.sorting.contract.IDisplaySortingUpdateEvent
 import kotlinx.coroutines.CoroutineScope
 
 interface ISortingAlgorithm
