+++ app/src/main/java/com/codingbot/algorithm/data/model/sorting/QuickSortAlgorithm.kt
@@ -1,8 +1,9 @@
-package com.codingbot.algorithm.data.model
+package com.codingbot.algorithm.data.model.sorting
 
 import com.codingbot.algorithm.core.common.Const
 import com.codingbot.algorithm.core.common.Logger
 import com.codingbot.algorithm.data.SortingData
+import com.codingbot.algorithm.data.model.sorting.contract.IDisplaySortingUpdateEvent
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
