+++ app/src/main/java/com/codingbot/algorithm/data/model/sorting/contract/IDisplayHeapSortingUpdateEvent.kt
@@ -1,4 +1,4 @@
-package com.codingbot.algorithm.data.model
+package com.codingbot.algorithm.data.model.sorting.contract
 
 import com.codingbot.algorithm.data.SortingData
