+++ app/src/main/java/com/codingbot/algorithm/viewmodel/SortingHeapSortingViewModel.kt
@@ -6,8 +6,8 @@ import com.codingbot.algorithm.core.common.Logger
 import com.codingbot.algorithm.core.common.SortingList
 import com.codingbot.algorithm.core.utils.scaledNumber
 import com.codingbot.algorithm.data.SortingData
-import com.codingbot.algorithm.data.model.HeapSortAlgorithm
-import com.codingbot.algorithm.data.model.IDisplayHeapSortingUpdateEvent
+import com.codingbot.algorithm.data.model.sorting.HeapSortAlgorithm
+import com.codingbot.algorithm.data.model.sorting.contract.IDisplayHeapSortingUpdateEvent
 import com.codingbot.algorithm.ui.ChannelUiEvent
 import com.codingbot.algorithm.ui.UiEvent
 import dagger.hilt.android.lifecycle.HiltViewModel
