+++ app/src/main/java/com/codingbot/algorithm/viewmodel/SortingViewModel.kt
@@ -6,12 +6,12 @@ import com.codingbot.algorithm.core.common.Logger
 import com.codingbot.algorithm.core.common.SortingList
 import com.codingbot.algorithm.core.utils.scaledNumber
 import com.codingbot.algorithm.data.SortingData
-import com.codingbot.algorithm.data.model.BubbleSortAlgorithm
-import com.codingbot.algorithm.data.model.IDisplaySortingUpdateEvent
-import com.codingbot.algorithm.data.model.ISortingAlgorithm
-import com.codingbot.algorithm.data.model.InsertionSortAlgorithm
-import com.codingbot.algorithm.data.model.QuickSortAlgorithm
-import com.codingbot.algorithm.data.model.SelectionSortAlgorithm
+import com.codingbot.algorithm.data.model.sorting.BubbleSortAlgorithm
+import com.codingbot.algorithm.data.model.sorting.contract.IDisplaySortingUpdateEvent
+import com.codingbot.algorithm.data.model.sorting.ISortingAlgorithm
+import com.codingbot.algorithm.data.model.sorting.InsertionSortAlgorithm
+import com.codingbot.algorithm.data.model.sorting.QuickSortAlgorithm
+import com.codingbot.algorithm.data.model.sorting.SelectionSortAlgorithm
 import com.codingbot.algorithm.ui.ChannelUiEvent
 import com.codingbot.algorithm.ui.UiEvent
 import dagger.hilt.android.lifecycle.HiltViewModel
